Refactor ReaderPage.kt into smaller components
- Extracted PostHeader components to PostHeader.kt
- Extracted MediaControls components to MediaControls.kt
- Improved readability and maintainability of ReaderPage.kt
- Removed unused imports and dead code in ReaderPage.kt
