+++ shared/src/commonMain/kotlin/dev/sasikanth/rss/reader/reader/page/ui/ReaderPage.kt
@@ -17,77 +17,24 @@
 
 package dev.sasikanth.rss.reader.reader.page.ui
 
-import androidx.compose.animation.AnimatedContent
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.LinearEasing
-import androidx.compose.animation.core.Spring
-import androidx.compose.animation.core.animateDpAsState
-import androidx.compose.animation.core.animateFloat
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.animation.core.infiniteRepeatable
-import androidx.compose.animation.core.rememberInfiniteTransition
-import androidx.compose.animation.core.spring
-import androidx.compose.animation.core.tween
-import androidx.compose.animation.expandIn
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.fadeIn
-import androidx.compose.animation.fadeOut
-import androidx.compose.animation.scaleIn
-import androidx.compose.animation.scaleOut
-import androidx.compose.animation.shrinkOut
-import androidx.compose.animation.shrinkVertically
-import androidx.compose.animation.slideInVertically
-import androidx.compose.animation.slideOutVertically
-import androidx.compose.animation.togetherWith
-import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.foundation.layout.requiredSize
-import androidx.compose.foundation.layout.requiredWidth
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.pager.PagerState
-import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.text.selection.DisableSelection
 import androidx.compose.foundation.text.selection.LocalTextSelectionColors
 import androidx.compose.foundation.text.selection.SelectionContainer
 import androidx.compose.foundation.text.selection.TextSelectionColors
-import androidx.compose.material3.CircularProgressIndicator
-import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
 import androidx.compose.material3.HorizontalDivider
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.LinearProgressIndicator
-import androidx.compose.material3.MaterialShapes
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
-import androidx.compose.material3.RadioButton
-import androidx.compose.material3.RadioButtonDefaults
-import androidx.compose.material3.Slider
-import androidx.compose.material3.SliderDefaults
-import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
-import androidx.compose.material3.TooltipAnchorPosition
-import androidx.compose.material3.TooltipBox
-import androidx.compose.material3.TooltipDefaults
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.material3.rememberTooltipState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.derivedStateOf
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
@@ -95,27 +42,10 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.BlendMode
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.hapticfeedback.HapticFeedbackType
-import androidx.compose.ui.platform.LocalHapticFeedback
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.semantics.clearAndSetSemantics
-import androidx.compose.ui.semantics.contentDescription
-import androidx.compose.ui.semantics.isTraversalGroup
-import androidx.compose.ui.semantics.role
-import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.TextLinkStyles
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextDecoration
-import androidx.compose.ui.text.style.TextOverflow
-import androidx.compose.ui.unit.IntSize
 import androidx.compose.ui.unit.dp
-import androidx.graphics.shapes.Morph
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.mikepenz.markdown.compose.LocalImageTransformer
 import com.mikepenz.markdown.compose.LocalMarkdownAnimations
@@ -126,7 +56,6 @@ import com.mikepenz.markdown.compose.LocalMarkdownPadding
 import com.mikepenz.markdown.compose.LocalMarkdownTypography
 import com.mikepenz.markdown.compose.LocalReferenceLinkHandler
 import com.mikepenz.markdown.compose.MarkdownElement
-import com.mikepenz.markdown.compose.components.CurrentComponentsBridge.text
 import com.mikepenz.markdown.compose.components.MarkdownComponents
 import com.mikepenz.markdown.m3.markdownColor
 import com.mikepenz.markdown.m3.markdownTypography
@@ -136,63 +65,18 @@ import com.mikepenz.markdown.model.markdownAnimations
 import com.mikepenz.markdown.model.markdownAnnotator
 import com.mikepenz.markdown.model.markdownDimens
 import com.mikepenz.markdown.model.markdownPadding
-import dev.sasikanth.rss.reader.components.image.FeedIcon
 import dev.sasikanth.rss.reader.core.model.local.ResolvedPost
-import dev.sasikanth.rss.reader.core.network.utils.UrlUtils
 import dev.sasikanth.rss.reader.data.repository.ReaderColorScheme
-import dev.sasikanth.rss.reader.home.ui.FeaturedImage
-import dev.sasikanth.rss.reader.home.ui.PostMetadataConfig
 import dev.sasikanth.rss.reader.markdown.CoilMarkdownTransformer
 import dev.sasikanth.rss.reader.media.PlaybackState
-import dev.sasikanth.rss.reader.media.SleepTimerOption
 import dev.sasikanth.rss.reader.platform.LocalLinkHandler
 import dev.sasikanth.rss.reader.reader.page.ReaderPageViewModel
 import dev.sasikanth.rss.reader.reader.page.ReaderProcessingProgress
-import dev.sasikanth.rss.reader.resources.icons.Bookmark
-import dev.sasikanth.rss.reader.resources.icons.Bookmarked
-import dev.sasikanth.rss.reader.resources.icons.Comments
-import dev.sasikanth.rss.reader.resources.icons.Forward30
-import dev.sasikanth.rss.reader.resources.icons.Pause
-import dev.sasikanth.rss.reader.resources.icons.Play
-import dev.sasikanth.rss.reader.resources.icons.Replay30
-import dev.sasikanth.rss.reader.resources.icons.Share
-import dev.sasikanth.rss.reader.resources.icons.Timer
-import dev.sasikanth.rss.reader.resources.icons.TwineIcons
-import dev.sasikanth.rss.reader.resources.icons.VisibilityOff
 import dev.sasikanth.rss.reader.share.LocalShareHandler
 import dev.sasikanth.rss.reader.ui.AppTheme
-import dev.sasikanth.rss.reader.ui.GolosFontFamily
-import dev.sasikanth.rss.reader.util.readerDateTimestamp
 import dev.sasikanth.rss.reader.utils.LocalBlockImage
-import dev.sasikanth.rss.reader.utils.ParallaxAlignment
-import dev.sasikanth.rss.reader.utils.getOffsetFractionForPage
-import dev.sasikanth.rss.reader.utils.toShape
 import kotlinx.coroutines.launch
-import kotlinx.serialization.json.Json
 import org.intellij.markdown.MarkdownElementTypes
-import org.jetbrains.compose.resources.stringResource
-import twine.shared.generated.resources.Res
-import twine.shared.generated.resources.bookmark
-import twine.shared.generated.resources.comments
-import twine.shared.generated.resources.markAsUnRead
-import twine.shared.generated.resources.pause
-import twine.shared.generated.resources.play
-import twine.shared.generated.resources.playback_speed
-import twine.shared.generated.resources.readingTimeEstimate
-import twine.shared.generated.resources.seek_backward
-import twine.shared.generated.resources.seek_forward
-import twine.shared.generated.resources.share
-import twine.shared.generated.resources.sleep_timer
-import twine.shared.generated.resources.sleep_timer_end_of_track
-import twine.shared.generated.resources.sleep_timer_minutes
-import twine.shared.generated.resources.sleep_timer_off
-import twine.shared.generated.resources.unBookmark
-
-private val json = Json {
-  ignoreUnknownKeys = true
-  isLenient = true
-  explicitNulls = false
-}
 
 @Composable
 internal fun ReaderPage(
@@ -388,708 +272,6 @@ private fun ProgressIndicator() {
     LinearProgressIndicator(
       trackColor = AppTheme.colorScheme.surfaceContainerLow,
       color = AppTheme.colorScheme.primary,
-    )
-  }
-}
-
-@Composable
-private fun PostHeader(
-  readerPost: ResolvedPost,
-  showFullArticle: Boolean,
-  page: Int,
-  pagerState: PagerState,
-  excerpt: String,
-  darkTheme: Boolean,
-  readerColorScheme: ReaderColorScheme,
-  onCommentsClick: () -> Unit,
-  onShareClick: () -> Unit,
-  onBookmarkClick: () -> Unit,
-  onMarkAsUnread: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  Column(modifier = Modifier.fillMaxWidth().then(modifier)) {
-    val title = readerPost.title
-    val description = readerPost.description
-    val postImage = readerPost.imageUrl
-
-    if (!postImage.isNullOrBlank()) {
-      Box(modifier = Modifier.padding(horizontal = 24.dp).align(Alignment.CenterHorizontally)) {
-        FeaturedImage(
-          imageUrl = postImage,
-          unlockAspectRatio = UrlUtils.isUnconstrainedMedia(postImage),
-          alignment =
-            remember(pagerState) {
-              ParallaxAlignment(
-                horizontalBias = { pagerState.getOffsetFractionForPage(page) },
-                multiplier = 2f,
-              )
-            },
-        )
-      }
-
-      Spacer(modifier = Modifier.requiredHeight(8.dp))
-    }
-
-    Column(modifier = Modifier.padding(horizontal = 32.dp)) {
-      DisableSelection {
-        Row(
-          modifier = Modifier.padding(top = 20.dp),
-          verticalAlignment = Alignment.CenterVertically,
-        ) {
-          Text(
-            text = readerPost.date.readerDateTimestamp(),
-            style = MaterialTheme.typography.bodyMedium,
-            color = AppTheme.colorScheme.outline,
-            maxLines = 1,
-          )
-
-          val readingTimeEstimate =
-            readerPost.articleContentReadingTime?.takeIf { showFullArticle }
-              ?: readerPost.feedContentReadingTime
-              ?: 0
-
-          if (readingTimeEstimate > 0) {
-            Text(
-              modifier = Modifier.padding(horizontal = 4.dp).clearAndSetSemantics {},
-              style = MaterialTheme.typography.bodyMedium,
-              maxLines = 1,
-              text = "\u2022",
-              color = AppTheme.colorScheme.outline,
-            )
-
-            Text(
-              text = stringResource(Res.string.readingTimeEstimate, readingTimeEstimate),
-              style = MaterialTheme.typography.bodyMedium,
-              color = AppTheme.colorScheme.outline,
-              maxLines = 1,
-            )
-          }
-        }
-      }
-
-      Text(
-        modifier =
-          Modifier.padding(top = 12.dp).graphicsLayer {
-            if (readerColorScheme == ReaderColorScheme.Dynamic) {
-              blendMode =
-                if (readerColorScheme.isDark(darkTheme)) {
-                  BlendMode.Screen
-                } else {
-                  BlendMode.Multiply
-                }
-            }
-          },
-        text = title.ifBlank { description },
-        style = MaterialTheme.typography.headlineMedium,
-        fontFamily = GolosFontFamily,
-        fontWeight = FontWeight.Bold,
-        color = AppTheme.colorScheme.secondary,
-        overflow = TextOverflow.Ellipsis,
-      )
-
-      if (excerpt.isNotBlank()) {
-        Spacer(Modifier.requiredHeight(8.dp))
-
-        Text(
-          text = excerpt,
-          style = MaterialTheme.typography.bodyMedium,
-          color = AppTheme.colorScheme.secondary,
-          maxLines = 3,
-          overflow = TextOverflow.Ellipsis,
-        )
-      }
-
-      Spacer(Modifier.requiredHeight(12.dp))
-
-      Row(verticalAlignment = Alignment.CenterVertically) {
-        DisableSelection {
-          PostSourcePill(
-            modifier = Modifier.weight(1f).clearAndSetSemantics {},
-            feedName = readerPost.feedName,
-            feedIcon = readerPost.feedIcon,
-            feedHomepageLink = readerPost.feedHomepageLink,
-            showFeedFavIcon = readerPost.showFeedFavIcon,
-            config =
-              PostMetadataConfig(
-                showUnreadIndicator = false,
-                showToggleReadUnreadOption = true,
-                enablePostSource = false,
-              ),
-            onSourceClick = {
-              // no-op
-            },
-          )
-        }
-
-        PostActions(
-          postBookmarked = readerPost.bookmarked,
-          commentsLink = readerPost.commentsLink,
-          onCommentsClick = onCommentsClick,
-          onShareClick = onShareClick,
-          onBookmarkClick = onBookmarkClick,
-          onMarkAsUnread = onMarkAsUnread,
-        )
-      }
-    }
-  }
-}
-
-@Composable
-private fun PostSourcePill(
-  feedIcon: String,
-  feedHomepageLink: String,
-  showFeedFavIcon: Boolean,
-  feedName: String,
-  config: PostMetadataConfig,
-  onSourceClick: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  Box(modifier = modifier) {
-    val postSourceTextColor =
-      if (config.enablePostSource) {
-        AppTheme.colorScheme.onSurface
-      } else {
-        AppTheme.colorScheme.onSurfaceVariant
-      }
-
-    Row(
-      modifier =
-        Modifier.background(
-            MaterialTheme.colorScheme.secondary.copy(alpha = 0.08f),
-            RoundedCornerShape(50),
-          )
-          .border(
-            1.dp,
-            MaterialTheme.colorScheme.secondary.copy(alpha = 0.16f),
-            androidx.compose.foundation.shape.RoundedCornerShape(50),
-          )
-          .clip(androidx.compose.foundation.shape.RoundedCornerShape(50))
-          .clickable(onClick = onSourceClick, enabled = config.enablePostSource)
-          .padding(vertical = 6.dp)
-          .padding(start = 8.dp, end = 12.dp),
-      verticalAlignment = Alignment.CenterVertically,
-    ) {
-      FeedIcon(
-        modifier = Modifier.requiredSize(16.dp),
-        icon = feedIcon,
-        homepageLink = feedHomepageLink,
-        showFeedFavIcon = showFeedFavIcon,
-        shape = MaterialTheme.shapes.extraSmall,
-        contentDescription = null,
-      )
-
-      Spacer(Modifier.requiredWidth(6.dp))
-
-      Text(
-        style = MaterialTheme.typography.labelMedium,
-        maxLines = 1,
-        text = feedName,
-        color = postSourceTextColor,
-        overflow = TextOverflow.Ellipsis,
-      )
-    }
-  }
-}
-
-@Composable
-private fun PostActions(
-  postBookmarked: Boolean,
-  commentsLink: String?,
-  onCommentsClick: () -> Unit,
-  onShareClick: () -> Unit,
-  onBookmarkClick: () -> Unit,
-  onMarkAsUnread: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  Row(modifier = modifier.semantics { isTraversalGroup = true }) {
-    val markAsUnreadLabel = stringResource(Res.string.markAsUnRead)
-    val markAsUnreadIcon = TwineIcons.VisibilityOff
-
-    PostActionButton(
-      label = markAsUnreadLabel,
-      icon = markAsUnreadIcon,
-      iconTint = AppTheme.colorScheme.onSurfaceVariant,
-      onClick = onMarkAsUnread,
-    )
-
-    if (!commentsLink.isNullOrBlank()) {
-      val commentsLabel = stringResource(Res.string.comments)
-      PostActionButton(
-        label = commentsLabel,
-        icon = TwineIcons.Comments,
-        iconTint = AppTheme.colorScheme.onSurfaceVariant,
-        onClick = onCommentsClick,
-      )
-    }
-
-    val sharedLabel = stringResource(Res.string.share)
-    PostActionButton(
-      label = sharedLabel,
-      icon = TwineIcons.Share,
-      iconTint = AppTheme.colorScheme.onSurfaceVariant,
-      onClick = onShareClick,
-    )
-
-    val bookmarkLabel =
-      if (postBookmarked) {
-        stringResource(Res.string.unBookmark)
-      } else {
-        stringResource(Res.string.bookmark)
-      }
-    PostActionButton(
-      label = bookmarkLabel,
-      icon =
-        if (postBookmarked) {
-          TwineIcons.Bookmarked
-        } else {
-          TwineIcons.Bookmark
-        },
-      iconTint =
-        if (postBookmarked) {
-          AppTheme.colorScheme.primary
-        } else {
-          AppTheme.colorScheme.onSurfaceVariant
-        },
-      onClick = onBookmarkClick,
-    )
-  }
-}
-
-@Composable
-private fun PostActionButton(
-  icon: ImageVector,
-  label: String,
-  modifier: Modifier = Modifier,
-  iconTint: Color = AppTheme.colorScheme.onSurface,
-  onClick: () -> Unit,
-) {
-  TooltipBox(
-    positionProvider =
-      TooltipDefaults.rememberTooltipPositionProvider(positioning = TooltipAnchorPosition.Above),
-    tooltip = {
-      Box(
-        modifier =
-          Modifier.background(AppTheme.colorScheme.surface, RoundedCornerShape(4.dp)).padding(8.dp)
-      ) {
-        Text(text = label)
-      }
-    },
-    state = rememberTooltipState(),
-  ) {
-    Box(
-      modifier =
-        Modifier.requiredSize(40.dp)
-          .clip(MaterialTheme.shapes.small)
-          .clickable(onClick = onClick)
-          .semantics {
-            role = Role.Button
-            contentDescription = label
-          }
-          .then(modifier),
-      contentAlignment = Alignment.Center,
-    ) {
-      Icon(
-        imageVector = icon,
-        contentDescription = null,
-        tint = iconTint,
-        modifier = Modifier.size(20.dp),
-      )
-    }
-  }
-}
-
-@OptIn(ExperimentalMaterial3ExpressiveApi::class)
-@Composable
-private fun MediaControls(
-  playbackState: PlaybackState,
-  onPlayClick: () -> Unit,
-  onPauseClick: () -> Unit,
-  onSeek: (Long) -> Unit,
-  onSeekForward: () -> Unit,
-  onSeekBackward: () -> Unit,
-  onPlaybackSpeedChange: (Float) -> Unit,
-  onSleepTimerClick: () -> Unit,
-  modifier: Modifier = Modifier,
-) {
-  val haptic = LocalHapticFeedback.current
-  val isPlaying = playbackState.isPlaying || playbackState.buffering
-  val progress =
-    if (playbackState.duration > 0) {
-      (playbackState.currentPosition.toFloat() / playbackState.duration.toFloat()).coerceIn(0f, 1f)
-    } else {
-      0f
-    }
-  val showExtendedControls = playbackState.duration > 0
-  val extendedControlsAnimationSpec =
-    spring<IntSize>(stiffness = Spring.StiffnessLow, dampingRatio = Spring.DampingRatioMediumBouncy)
-
-  Column(
-    modifier =
-      modifier
-        .fillMaxWidth()
-        .background(AppTheme.colorScheme.surface, RoundedCornerShape(16.dp))
-        .padding(16.dp),
-    verticalArrangement = Arrangement.spacedBy(if (showExtendedControls) 16.dp else 0.dp),
-  ) {
-    AnimatedVisibility(
-      visible = showExtendedControls,
-      enter = fadeIn() + expandVertically(animationSpec = extendedControlsAnimationSpec),
-      exit = fadeOut() + shrinkVertically(animationSpec = extendedControlsAnimationSpec),
-    ) {
-      Column {
-        val sliderColors =
-          SliderDefaults.colors(
-            activeTrackColor = AppTheme.colorScheme.primaryContainer,
-            inactiveTrackColor = AppTheme.colorScheme.surfaceContainerHigh,
-          )
-        Slider(
-          modifier = Modifier.padding(top = 8.dp),
-          value = progress,
-          onValueChange = {
-            val newPosition = (it * playbackState.duration).toLong()
-            val currentPosition = (progress * playbackState.duration).toLong()
-            if (newPosition / 1000 != currentPosition / 1000) {
-              haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-            }
-            onSeek(newPosition)
-          },
-          thumb = {
-            Box(
-              modifier =
-                Modifier.requiredSize(24.dp)
-                  .shadow(elevation = 2.dp, shape = CircleShape)
-                  .background(AppTheme.colorScheme.inverseSurface, CircleShape)
-                  .border(1.dp, AppTheme.colorScheme.secondary, CircleShape)
-            )
-          },
-          track = {
-            SliderDefaults.Track(sliderState = it, thumbTrackGapSize = 0.dp, colors = sliderColors)
-          },
-        )
-
-        Row(modifier = Modifier.fillMaxWidth()) {
-          Text(
-            text = formatDuration(playbackState.currentPosition),
-            style = MaterialTheme.typography.labelSmall,
-            color = AppTheme.colorScheme.onSurfaceVariant,
-          )
-          Spacer(Modifier.weight(1f))
-          Text(
-            text = formatDuration(playbackState.duration),
-            style = MaterialTheme.typography.labelSmall,
-            color = AppTheme.colorScheme.onSurfaceVariant,
-          )
-        }
-      }
-    }
-
-    Row(
-      modifier = Modifier.fillMaxWidth(),
-      verticalAlignment = Alignment.CenterVertically,
-      horizontalArrangement = Arrangement.SpaceEvenly,
-    ) {
-      AnimatedVisibility(
-        visible = showExtendedControls,
-        enter =
-          fadeIn() +
-            expandIn(
-              animationSpec = extendedControlsAnimationSpec,
-              expandFrom = Alignment.Center,
-              clip = false,
-            ),
-        exit =
-          fadeOut() +
-            shrinkOut(
-              animationSpec = extendedControlsAnimationSpec,
-              shrinkTowards = Alignment.Center,
-              clip = false,
-            ),
-      ) {
-        TextButton(
-          onClick = {
-            haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-            onPlaybackSpeedChange(playbackState.playbackSpeed)
-          }
-        ) {
-          AnimatedContent(
-            targetState = playbackState.playbackSpeed,
-            transitionSpec = {
-              (fadeIn() + scaleIn() + slideInVertically()).togetherWith(
-                (fadeOut() + scaleOut() + slideOutVertically { it / 2 })
-              )
-            },
-          ) {
-            Text(
-              text = stringResource(Res.string.playback_speed, it),
-              style = MaterialTheme.typography.labelLarge,
-              color = AppTheme.colorScheme.onSurfaceVariant,
-            )
-          }
-        }
-      }
-
-      AnimatedVisibility(
-        visible = showExtendedControls,
-        enter =
-          fadeIn() +
-            expandIn(
-              animationSpec = extendedControlsAnimationSpec,
-              expandFrom = Alignment.Center,
-              clip = false,
-            ),
-        exit =
-          fadeOut() +
-            shrinkOut(
-              animationSpec = extendedControlsAnimationSpec,
-              shrinkTowards = Alignment.Center,
-              clip = false,
-            ),
-      ) {
-        IconButton(
-          onClick = {
-            haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-            onSeekBackward()
-          }
-        ) {
-          Icon(
-            imageVector = TwineIcons.Replay30,
-            contentDescription = stringResource(Res.string.seek_backward),
-            tint = AppTheme.colorScheme.onSurfaceVariant,
-          )
-        }
-      }
-
-      val progress by animateFloatAsState(if (isPlaying) 1f else 0f)
-      val buttonSize by
-        animateDpAsState(
-          if (isPlaying) {
-            64.dp
-          } else {
-            56.dp
-          },
-          animationSpec =
-            spring(
-              stiffness = Spring.StiffnessMedium,
-              dampingRatio = Spring.DampingRatioMediumBouncy,
-            ),
-        )
-      val fabMorph by remember {
-        derivedStateOf { Morph(start = MaterialShapes.Circle, end = MaterialShapes.Cookie9Sided) }
-      }
-      val infiniteTransition = rememberInfiniteTransition(label = "playButtonRotation")
-      val rotationAngle by
-        infiniteTransition.animateFloat(
-          initialValue = 0f,
-          targetValue = 360f,
-          animationSpec =
-            infiniteRepeatable(animation = tween(durationMillis = 6000, easing = LinearEasing)),
-          label = "playButtonRotatingAngle",
-        )
-
-      Box(contentAlignment = Alignment.Center, modifier = Modifier.size(64.dp)) {
-        // Rotating shape background
-        Box(
-          modifier =
-            Modifier.size(buttonSize)
-              .graphicsLayer { rotationZ = if (isPlaying) rotationAngle else 0f }
-              .background(AppTheme.colorScheme.primaryContainer, fabMorph.toShape(progress))
-        )
-
-        // Non-rotating icon
-        Box(
-          modifier =
-            Modifier.matchParentSize()
-              .clip(fabMorph.toShape(progress))
-              .clickable(
-                interactionSource = remember { MutableInteractionSource() },
-                indication = null,
-                onClick = {
-                  haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-                  if (isPlaying) onPauseClick() else onPlayClick()
-                },
-              ),
-          contentAlignment = Alignment.Center,
-        ) {
-          if (playbackState.buffering) {
-            CircularProgressIndicator(
-              modifier = Modifier.size(32.dp),
-              color = AppTheme.colorScheme.onPrimaryContainer,
-              strokeWidth = 2.dp,
-            )
-          } else {
-            Icon(
-              modifier = Modifier.size(32.dp),
-              imageVector = if (isPlaying) TwineIcons.Pause else TwineIcons.Play,
-              contentDescription =
-                if (isPlaying) stringResource(Res.string.pause)
-                else stringResource(Res.string.play),
-              tint = AppTheme.colorScheme.onPrimaryContainer,
-            )
-          }
-        }
-      }
-
-      AnimatedVisibility(
-        visible = showExtendedControls,
-        enter =
-          fadeIn() +
-            expandIn(
-              animationSpec = extendedControlsAnimationSpec,
-              expandFrom = Alignment.Center,
-              clip = false,
-            ),
-        exit =
-          fadeOut() +
-            shrinkOut(
-              animationSpec = extendedControlsAnimationSpec,
-              shrinkTowards = Alignment.Center,
-              clip = false,
-            ),
-      ) {
-        IconButton(
-          onClick = {
-            haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-            onSeekForward()
-          }
-        ) {
-          Icon(
-            imageVector = TwineIcons.Forward30,
-            contentDescription = stringResource(Res.string.seek_forward),
-            tint = AppTheme.colorScheme.onSurfaceVariant,
-          )
-        }
-      }
-
-      AnimatedVisibility(
-        visible = showExtendedControls,
-        enter =
-          fadeIn() +
-            expandIn(
-              animationSpec = extendedControlsAnimationSpec,
-              expandFrom = Alignment.Center,
-              clip = false,
-            ),
-        exit =
-          fadeOut() +
-            shrinkOut(
-              animationSpec = extendedControlsAnimationSpec,
-              shrinkTowards = Alignment.Center,
-              clip = false,
-            ),
-      ) {
-        IconButton(
-          onClick = {
-            haptic.performHapticFeedback(HapticFeedbackType.TextHandleMove)
-            onSleepTimerClick()
-          }
-        ) {
-          val sleepTimerRemaining = playbackState.sleepTimerRemaining
-          val text =
-            if (sleepTimerRemaining != null && sleepTimerRemaining > 0) {
-              formatDuration(sleepTimerRemaining)
-            } else if (sleepTimerRemaining == -1L) {
-              stringResource(Res.string.sleep_timer_end_of_track)
-            } else {
-              stringResource(Res.string.sleep_timer)
-            }
-
-          Icon(
-            imageVector = TwineIcons.Timer,
-            contentDescription = text,
-            tint = AppTheme.colorScheme.onSurfaceVariant,
-          )
-        }
-      }
-    }
-  }
-}
-
-private fun formatDuration(duration: Long): String {
-  val seconds = (duration / 1000) % 60
-  val minutes = (duration / (1000 * 60)) % 60
-  val hours = (duration / (1000 * 60 * 60))
-
-  return if (hours > 0) {
-    "${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}"
-  } else {
-    "${minutes.toString().padStart(2, '0')}:${seconds.toString().padStart(2, '0')}"
-  }
-}
-
-@Composable
-private fun SleepTimerBottomSheet(
-  playbackState: PlaybackState,
-  onOptionSelected: (SleepTimerOption) -> Unit,
-  onDismiss: () -> Unit,
-) {
-  ModalBottomSheet(
-    onDismissRequest = onDismiss,
-    containerColor = AppTheme.colorScheme.surfaceContainerLowest,
-    contentColor = AppTheme.colorScheme.onSurface,
-    sheetState = rememberModalBottomSheetState(skipPartiallyExpanded = true),
-  ) {
-    Column(
-      modifier =
-        Modifier.fillMaxWidth().padding(horizontal = 24.dp).padding(top = 8.dp, bottom = 48.dp),
-      verticalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-      Text(
-        text = stringResource(Res.string.sleep_timer),
-        style = MaterialTheme.typography.titleLarge,
-        modifier = Modifier.padding(bottom = 16.dp),
-      )
-
-      val currentOption = playbackState.selectedSleepTimerOption
-
-      SleepTimerOptionItem(
-        label = stringResource(Res.string.sleep_timer_off),
-        isSelected = currentOption is SleepTimerOption.None,
-        onClick = { onOptionSelected(SleepTimerOption.None) },
-      )
-
-      SleepTimerOptionItem(
-        label = stringResource(Res.string.sleep_timer_end_of_track),
-        isSelected = currentOption is SleepTimerOption.EndOfTrack,
-        onClick = { onOptionSelected(SleepTimerOption.EndOfTrack) },
-      )
-
-      listOf(15, 30, 45, 60).forEach { minutes ->
-        val option = SleepTimerOption.Minutes(minutes)
-        SleepTimerOptionItem(
-          label = stringResource(Res.string.sleep_timer_minutes, minutes),
-          isSelected = currentOption == option,
-          onClick = { onOptionSelected(option) },
-        )
-      }
-    }
-  }
-}
-
-@Composable
-private fun SleepTimerOptionItem(label: String, isSelected: Boolean, onClick: () -> Unit) {
-  Row(
-    modifier =
-      Modifier.fillMaxWidth()
-        .clip(RoundedCornerShape(12.dp))
-        .clickable(onClick = onClick)
-        .padding(vertical = 12.dp),
-    verticalAlignment = Alignment.CenterVertically,
-    horizontalArrangement = Arrangement.spacedBy(16.dp),
-  ) {
-    RadioButton(
-      selected = isSelected,
-      onClick = null,
-      colors =
-        RadioButtonDefaults.colors(
-          selectedColor = AppTheme.colorScheme.primary,
-          unselectedColor = AppTheme.colorScheme.outline,
-        ),
-    )
-
-    Text(
-      text = label,
-      style = MaterialTheme.typography.bodyLarge,
-      color = if (isSelected) AppTheme.colorScheme.primary else AppTheme.colorScheme.onSurface,
     )
   }
 }
